% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_SCOPE.R
\name{run_scope}
\alias{run_scope}
\title{Runs SCOPE.m in R}
\usage{
run_scope(SCOPE_path = NULL)
}
\arguments{
\item{SCOPE_path}{filepath where SCOPE.m is located}
}
\value{
The function accepts a string parameter, path, that represents the directory path where SCOPE.m is located. If the path is not specified, the function prompts the user to choose the file location interactively. Once the path is determined, the function saves it to an R environment variable named SCOPE_PATH, which can be used in subsequent function calls to avoid re-locating SCOPE.m.

Additionally, the function creates an R session hook that removes the SCOPE_PATH variable when the R session is terminated, ensuring that the path is not persisted across different sessions.

This function can be called with a directory path as an argument, or without arguments to prompt the user to select the file interactively. Once the path is determined, the function saves it to the SCOPE_PATH variable in the global environment. The on.exit() hook ensures that the SCOPE_PATH variable is removed when the R session is terminated.
}
\description{
This function runs SCOPE.m in R language.
}
