% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_set_filenames.R
\name{set_filenames}
\alias{set_filenames}
\title{Create and/or modify datasheet}
\usage{
set_filenames(
  simulation_name = "verification_run",
  dataset_dir = "for_verification",
  meteo_ec_csv = "input_data_latin_hypercube_ts",
  t = TRUE,
  Rin = TRUE,
  Rli = FALSE,
  p = FALSE,
  Ta = TRUE,
  u = FALSE,
  ea = FALSE,
  RH = TRUE,
  VPD = FALSE,
  tts = TRUE,
  tto = TRUE,
  psi = TRUE,
  Cab = TRUE,
  Cca = TRUE,
  Cdm = TRUE,
  Cw = TRUE,
  Cs = TRUE,
  Cant = TRUE,
  SMC = FALSE,
  BSMBrightness = FALSE,
  BSMlat = FALSE,
  BSMlon = FALSE,
  LAI = TRUE,
  hc = FALSE,
  LIDFa = TRUE,
  LIDFb = TRUE,
  z = FALSE,
  Ca = FALSE,
  Vcmax25 = TRUE,
  BallBerrySlope = TRUE,
  interactive = FALSE,
  rows_to_edit = NULL
)
}
\arguments{
\item{interactive}{If TRUE, user can edit rows interactively}

\item{rows_to_edit}{Vector of row indices to edit (only if interactive = TRUE)}
}
\value{
A data frame with two columns (col1, col2) representing the datasheet
}
\description{
This function generates the datasheet from a fixed template
and lets you either:
\itemize{
\item Automatically replace values using arguments (like set_filenames)
\item Interactively edit rows in the second column
}
}
